Remove external function reference in constructor.
The present ERC165Mapping implementation is from the EIP 165 spec at:

  https://github.com/ethereum/EIPs/blob/master/EIPS/eip-165.md

and in particular, the line:

  supportedInterfaces[this.supportsInterface.selector] = true;

is used to set the value of 'supportedInterfaces' at the given key to
true.  The key itself is the Ethereum ABI "function selector" for the
'supportsInterface' function, the spec of which is detailed here:

  https://solidity.readthedocs.io/en/v0.4.24/abi-spec.html#function-selector

From the spec: the function selector is the first (left, high-order in
big endian) four bytes of the Keccak (SHA-3) hash of the "signature" of
the function, where the signature is the function name along with a
parenthesised list of parameter types (sans spaces).

The function selector being calculated above is thus:

  bytes4(keccak256('supportsInterface(bytes4)'))

which evaluates to 0x01ffc9a7, as also pointed out in the EIP 165 spec.
In this case it's used to identify an interface uniquely; an interface
identifier is the XOR of all function selectors in the interface, which
for ERC165 is just the single one corresponding to 'supportsInterface'.

In the sample ERC165MappingImplementation given in the EIP 165 spec, the
authors use 'this.supportsInterface.selector' instead of the hard-coded
selector address 0x01ffc9a7.  This yields a compiler warning:
'supportsInterface' is an external function and thus can't be called
while the contract is being constructed, so any reference to it there
(via 'this') triggers a warning.  The function is not actually *called*
here -- just referred to -- but the warning persists nonetheless, and
warnings suck.

Following the pattern used for registering supported interfaces in
Constitution.sol, I've simply replaced the warning-causing
'this.supportsInterface.selector' expression here with the hard-coded
interface identifier, 0x01ffc9a7.